@@ -1,6 +1,6 @@
 {
-	"name": "app",
-	"version": "0.0.1",
+	"name": "virtual-tour-svelte-app",
+	"version": "1.0.0",
 	"private": true,
 	"scripts": {
 		"dev": "vite dev",
@@ -8,8 +8,9 @@
 		"preview": "vite preview",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write ."
+		"lint": "prettier --ignore-path .gitignore --check --plugin-search-dir=. . && eslint .",
+		"lint:fix": "eslint . --fix",
+		"format": "prettier --ignore-path .gitignore --write --plugin-search-dir=. ."
 	},
 	"dependencies": {
 		"@photo-sphere-viewer/core": "^5.4.2",
@@ -17,28 +18,33 @@
 		"@photo-sphere-viewer/virtual-tour-plugin": "^5.4.2",
 		"@sanity/client": "^6.0.0",
 		"@sanity/image-url": "^1.0.2",
-		"groq": "^3.8.1"
+		"@tailwindcss/vite": "^4.0.0-beta.6",
+		"groq": "^3.8.1",
+		"typescript-eslint": "^8.18.1"
 	},
 	"devDependencies": {
-		"@portabletext/svelte": "^2.0.0",
+		"@eslint/compat": "^1.2.4",
 		"@portabletext/types": "^2.0.2",
 		"@sanity/types": "^3.7.1",
-		"@sveltejs/adapter-auto": "^2.0.0",
-		"@sveltejs/kit": "^1.5.0",
-		"@typescript-eslint/eslint-plugin": "^5.45.0",
-		"@typescript-eslint/parser": "^5.45.0",
-		"autoprefixer": "^10.4.14",
-		"eslint": "^8.28.0",
-		"eslint-config-prettier": "^8.5.0",
-		"eslint-plugin-svelte3": "^4.0.0",
-		"postcss": "^8.4.22",
-		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.8.1",
-		"svelte": "^3.54.0",
-		"svelte-check": "^3.0.1",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.10.1",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@typescript-eslint/eslint-plugin": "^8.18.1",
+		"@typescript-eslint/parser": "^8.18.1",
+		"autoprefixer": "^10.4.20",
+		"eslint": "^9.17.0",
+		"eslint-config-prettier": "^9.0.0",
+		"eslint-plugin-svelte": "^2.34.0",
+		"postcss": "^8.4.49",
+		"prettier": "^3.0.3",
+		"prettier-plugin-svelte": "^3.0.3",
+		"svelte": "^5.15.0",
+		"svelte-check": "^4.1.1",
+		"tailwindcss": "^3.4.17",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
-		"vite": "^4.2.0"
+		"vite": "^6.0.5",
+		"vitest": "^2.1.8"
 	},
 	"type": "module",
 	"engines": {
