@@ -1,7 +1,7 @@
 {
-  "name": "app",
+  "name": "virtual-tour-studio",
   "private": true,
-  "version": "1.0.0",
+  "version": "1.1.1",
   "main": "package.json",
   "license": "UNLICENSED",
   "scripts": {
@@ -9,26 +9,26 @@
     "start": "sanity start",
     "build": "sanity build",
     "deploy": "sanity deploy",
-    "deploy-graphql": "sanity graphql deploy"
+    "deploy-graphql": "sanity graphql deploy",
+    "format": "pnpx prettier --write \"**/*.{md,css,js,ts,json}\" \"!node_modules/**\"",
+    "lint": "eslint ."
   },
   "keywords": [
     "sanity"
   ],
   "dependencies": {
     "@sanity/vision": "^3.8.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-is": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "sanity": "^3.8.1",
-    "styled-components": "^5.2.0"
+    "styled-components": "^6.1.13"
   },
   "devDependencies": {
-    "@sanity/eslint-config-studio": "^2.0.1",
+    "@sanity/eslint-config-studio": "^5.0.1",
     "@types/node": "^18.15.11",
     "@types/react": "^18.0.25",
-    "@types/styled-components": "^5.1.26",
-    "eslint": "^8.6.0",
-    "prettier": "^2.8.4",
+    "eslint": "^9.14.0",
+    "prettier": "^3.0.3",
     "typescript": "^5.0.2"
   },
   "prettier": {
