@@ -1,3 +1,8 @@
-<div class="app">
-  <slot />
-</div>
+<script>
+	import '$lib/styles/style.css';
+	let { children } = $props();
+</script>
+
+<main class="container mx-auto py-8">
+	{@render children()}
+</main>
