@@ -1,27 +1,31 @@
-<script>
-  import Virtual from '$lib/Virtual.svelte'
+<script lang="ts">
+	import Virtual from '$lib/components/Virtual.svelte';
+	import type { PageData } from './$types';
 
-  export let data
+	let { data }: { data: PageData } = $props();
 
-  let e = data.e
+	let e = data.e;
 </script>
 
 <svelte:head>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <meta property="og:title" content={e.virtualTourPageBlocks.title} />
-  <meta property="og:description" content={e.virtualTourPageBlocks.description} />
-  <meta property="og:type" content="website" />
+	<title>{e.virtualTourPageBlocks.title}</title>
+	<meta name="description" content={e.virtualTourPageBlocks.description} />
+	<meta property="og:title" content={e.virtualTourPageBlocks.title} />
+	<meta property="og:description" content={e.virtualTourPageBlocks.description} />
+	<meta property="og:type" content="website" />
 </svelte:head>
 
+<h1>{e.virtualTourPageBlocks.title}</h1>
+<p>{e.virtualTourPageBlocks.description}</p>
 <div class="app">
-  <Virtual virtualTourPageBlocks={e.virtualTourPageBlocks} virtualTourItem={e.virtualTourItem}/>
+	<Virtual virtualTourPageBlocks={e.virtualTourPageBlocks} virtualTourItem={e.virtualTourItem} />
 </div>
 
 <style>
-  .app {
-    margin: 0;
-    width: 100vw;
-    height: 100vh;
-    font-family: sans-serif;
-  }
+	.app {
+		margin: 0;
+		width: 80vw;
+		height: 80vh;
+		font-family: sans-serif;
+	}
 </style>
