@@ -2,7 +2,7 @@ import {visionTool} from '@sanity/vision'
 import {defineConfig} from 'sanity'
 import {deskTool} from 'sanity/desk'
 import {schemaTypes} from './schemas'
-import { structure } from './deskStructure'
+import {structure} from './deskStructure'
 
 export const projectId = process.env.SANITY_STUDIO_PROJECT_ID!
 export const dataset = process.env.SANITY_STUDIO_DATASET!
@@ -12,9 +12,12 @@ export default defineConfig({
   title: 'Project Name',
   projectId,
   dataset,
-  plugins: [deskTool({
-    structure
-  }), visionTool()],
+  plugins: [
+    deskTool({
+      structure,
+    }),
+    visionTool(),
+  ],
   schema: {
     types: schemaTypes,
   },
